requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
huggingface-hub>=0.19.4
requests>=2.31.0